people_counting/requirements.txt
numpy
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort
imageio
tensorflow==1.15
tensorflow-gpu==1.15
scikit-learn==0.22.2
#keras==2.4.0
keras==2.3.1


# The following are added by ashish
pandas-gbq
google-cloud-bigquery
